+++ src/class/templates/admin/data/admin-data-person-display.php
@@ -26,7 +26,7 @@ $lumiere_comments_fields = Get_Options_Person::get_items_person_details_comments
 
 	<?php
 	// Common menu for data
-	require_once __DIR__ . '/admin-data-person-first-menu.php';
+	require_once __DIR__ . '/admin-data-person-submenu.php';
 	?>
 
 	<div class="lumiere_title_options lumiere_border_shadow">