+++ tests/codeception/Acceptance/013-SwitchModalWindowsCest.php
@@ -51,7 +51,7 @@ class ModalWindowsCest {
 		$I->seeInPageSource("lumiere_style_main-css"); 	
 		$I->seeInPageSource("lumiere_highslide_core-js");
 		$I->seeInPageSource("lumiere_highslide_options-js");
-		$I->seeInPageSource('<a class="add_cursor lum_link_make_popup lum_link_with_people highslide" id="link-0227759" data-modal_window_nonce="' );
+		$I->seeInPageSource('<a class="lum_link add_cursor lum_link_make_popup lum_link_with_people highslide" id="link-0227759" data-modal_window_nonce="' );
 		$I->seeInPageSource( 'data-modal_window_people="0227759" data-target="#theModal0227759" title="Open a new window with IMDb informations for Peter Dinklage">Peter Dinklage</a>' );
 		$I->amOnPage( AcceptanceSettings::TESTING_PAGE_TAXONOMY_URL );# Check taxonomy page
 		$I->seeInPageSource("lumiere_highslide_core_style-css");
@@ -75,7 +75,7 @@ class ModalWindowsCest {
 		
 		$I->amOnPage( AcceptanceSettings::TESTING_PAGE_BASE_URL );# Check regular page
 		$I->seeInPageSource("lumiere_bootstrap_custom-css");		
-		$I->seeInPageSource( '<a class="add_cursor lum_link_make_popup lum_link_with_people" id="link-0227759" data-modal_window_nonce="' );
+		$I->seeInPageSource( '<a class="lum_link add_cursor lum_link_make_popup lum_link_with_people" id="link-0227759" data-modal_window_nonce="' );
 		$I->seeInPageSource( 'data-modal_window_people="0227759" data-target="#theModal0227759" title="Open a new window with IMDb informations for Peter Dinklage">Peter Dinklage</a>');
 		$I->amOnPage( AcceptanceSettings::TESTING_PAGE_TAXONOMY_URL );# Check taxonomy page
 		$I->seeInPageSource("lumiere_bootstrap_custom-css");			
@@ -93,7 +93,7 @@ class ModalWindowsCest {
 		$I->waitPageLoad();
 		$I->amOnPage( AcceptanceSettings::TESTING_PAGE_BASE_URL );# Check regular page
 		$I->seeInPageSource("lumiere_classic_links-js");		
-		$I->seeInPageSource( '<a class="add_cursor lum_link_make_popup lum_link_with_people" id="link-0227759" data-modal_window_nonce="' );
+		$I->seeInPageSource( '<a class="lum_link add_cursor lum_link_make_popup lum_link_with_people" id="link-0227759" data-modal_window_nonce="' );
 		$I->seeInPageSource( 'data-modal_window_people="0227759" data-target="#theModal0227759" title="Open a new window with IMDb informations for Peter Dinklage">Peter Dinklage</a>');
 		$I->amOnPage( AcceptanceSettings::TESTING_PAGE_TAXONOMY_URL );# Check taxonomy page
 		$I->seeInPageSource("lumiere_classic_links-js");
@@ -111,7 +111,7 @@ class ModalWindowsCest {
 		$I->amOnPage( AcceptanceSettings::TESTING_PAGE_BASE_URL . '?amp' );# Check regular page
 		$I->waitForText( 'test codeception', 15 ); // wait up to 15 seconds
 		$I->seeInPageSource("<link rel=\"preconnect\" href=\"https://cdn.ampproject.org\">");
-		$I->seeInPageSource('<a class="add_cursor lum_link_no_popup" id="link-0227759" data-modal_window_nonce="');
+		$I->seeInPageSource('<a class="lum_link add_cursor lum_link_no_popup" id="link-0227759" data-modal_window_nonce="');
 		$I->seeInPageSource('data-modal_window_people="0227759" data-target="#theModal0227759" title="Open a new window with IMDb informations for Peter Dinklage" href="' . $I->getCustomBaseUrl() . '/lumiere/person/?mid=0227759&amp;');
 		$I->seeInPageSource( '&amp;amp">Peter Dinklage</a>');
 		$I->amOnPage( AcceptanceSettings::TESTING_PAGE_TAXONOMY_URL . '?amp' );# Check taxonomy page